Simplify PR for review
