+++ src/cascadia/TerminalSettingsEditor/CommonResources.xaml
@@ -8,14 +8,27 @@
                     xmlns:mtu="using:Microsoft.Terminal.UI"
                     xmlns:muxc="using:Microsoft.UI.Xaml.Controls">
 
-    <!--  Merge SettingsControls and SettingContainer styles here to give every page access to them  -->
+    <!--  Merge SettingsControls styles here to give every page access to them  -->
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="SettingsControlsStyle.xaml" />
-        <ResourceDictionary Source="SettingContainerStyle.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
     <ResourceDictionary.ThemeDictionaries>
         <ResourceDictionary x:Key="Dark">
+            <Style x:Key="SecondaryTextBlockStyle"
+                   TargetType="TextBlock">
+                <Setter Property="Foreground" Value="{ThemeResource TextFillColorSecondary}" />
+            </Style>
+            <SolidColorBrush x:Key="SubgroupHeaderBrush"
+                             Color="{StaticResource TextFillColorSecondary}" />
+            <StaticResource x:Key="SettingContainerErrorSeverityBackgroundBrush"
+                            ResourceKey="SystemFillColorCriticalBackgroundBrush" />
+            <StaticResource x:Key="SettingContainerWarningSeverityBackgroundBrush"
+                            ResourceKey="SystemFillColorCautionBackgroundBrush" />
+            <StaticResource x:Key="SettingContainerErrorSeverityIconBackground"
+                            ResourceKey="SystemFillColorCriticalBrush" />
+            <StaticResource x:Key="SettingContainerWarningSeverityIconBackground"
+                            ResourceKey="SystemFillColorCautionBrush" />
             <SolidColorBrush x:Key="DeleteButtonBackground"
                              Color="{ThemeResource DeleteButtonColor}" />
             <SolidColorBrush x:Key="DeleteButtonBackgroundPointerOver"
@@ -32,6 +45,20 @@
                              Color="#B3FFFFFF" />
         </ResourceDictionary>
         <ResourceDictionary x:Key="Light">
+            <Style x:Key="SecondaryTextBlockStyle"
+                   TargetType="TextBlock">
+                <Setter Property="Foreground" Value="{ThemeResource TextFillColorSecondary}" />
+            </Style>
+            <SolidColorBrush x:Key="SubgroupHeaderBrush"
+                             Color="{StaticResource TextFillColorSecondary}" />
+            <StaticResource x:Key="SettingContainerErrorSeverityBackgroundBrush"
+                            ResourceKey="SystemFillColorCriticalBackgroundBrush" />
+            <StaticResource x:Key="SettingContainerWarningSeverityBackgroundBrush"
+                            ResourceKey="SystemFillColorCautionBackgroundBrush" />
+            <StaticResource x:Key="SettingContainerErrorSeverityIconBackground"
+                            ResourceKey="SystemFillColorCriticalBrush" />
+            <StaticResource x:Key="SettingContainerWarningSeverityIconBackground"
+                            ResourceKey="SystemFillColorCautionBrush" />
             <SolidColorBrush x:Key="DeleteButtonBackground"
                              Color="{ThemeResource DeleteButtonColor}" />
             <SolidColorBrush x:Key="DeleteButtonBackgroundPointerOver"
@@ -48,6 +75,18 @@
                              Color="#B3FFFFFF" />
         </ResourceDictionary>
         <ResourceDictionary x:Key="HighContrast">
+            <Style x:Key="SecondaryTextBlockStyle"
+                   TargetType="TextBlock" />
+            <SolidColorBrush x:Key="SubgroupHeaderBrush"
+                             Color="{ThemeResource SystemColorWindowTextColor}" />
+            <StaticResource x:Key="SettingContainerErrorSeverityBackgroundBrush"
+                            ResourceKey="SystemColorWindowColorBrush" />
+            <StaticResource x:Key="SettingContainerWarningSeverityBackgroundBrush"
+                            ResourceKey="SystemColorWindowColorBrush" />
+            <StaticResource x:Key="SettingContainerErrorSeverityIconBackground"
+                            ResourceKey="SystemColorHighlightColorBrush" />
+            <StaticResource x:Key="SettingContainerWarningSeverityIconBackground"
+                            ResourceKey="SystemColorHighlightColorBrush" />
             <SolidColorBrush x:Key="DeleteButtonBackground"
                              Color="{ThemeResource SystemColorButtonFaceColor}" />
             <SolidColorBrush x:Key="DeleteButtonBackgroundPointerOver"
@@ -62,6 +101,31 @@
                              Color="{ThemeResource SystemColorHighlightTextColor}" />
         </ResourceDictionary>
     </ResourceDictionary.ThemeDictionaries>
+
+    <!--  Glyphs / sizing for the inline error & warning severity badges (e.g. on Appearances).  -->
+    <x:String x:Key="SettingContainerErrorIconGlyph">&#xF13D;</x:String>
+    <x:String x:Key="SettingContainerWarningIconGlyph">&#xF13C;</x:String>
+    <x:Double x:Key="SettingContainerIconFontSize">16</x:Double>
+
+    <Style x:Key="SettingsPageItemDescriptionStyle"
+           BasedOn="{StaticResource CaptionTextBlockStyle}"
+           TargetType="TextBlock">
+        <Setter Property="Margin" Value="0,0,24,0" />
+        <Setter Property="LineHeight" Value="16" />
+        <Setter Property="Foreground" Value="{ThemeResource SubgroupHeaderBrush}" />
+        <Setter Property="TextWrapping" Value="WrapWholeWords" />
+        <Setter Property="FontFamily" Value="Segoe UI, Segoe Fluent Icons, Segoe MDL2 Assets" />
+    </Style>
+
+    <Style x:Key="SettingContainerCurrentValueTextBlockStyle"
+           BasedOn="{StaticResource SettingsPageItemDescriptionStyle}"
+           TargetType="TextBlock">
+        <Setter Property="MaxWidth" Value="248" />
+        <Setter Property="Margin" Value="0" />
+        <Setter Property="HorizontalAlignment" Value="Right" />
+        <Setter Property="VerticalAlignment" Value="Center" />
+        <Setter Property="FontFamily" Value="Segoe UI, Segoe Fluent Icons, Segoe MDL2 Assets" />
+    </Style>
 
     <local:ColorToBrushConverter x:Key="ColorToBrushConverter" />
     <local:ColorToStringConverter x:Key="ColorToStringConverter" />
