+++ src/cascadia/TerminalSettingsEditor/ProfileViewModel.h
