+++ src/cascadia/TerminalSettingsEditor/ProfileViewModel.idl
