--- src/cascadia/TerminalSettingsEditor/SettingContainerStyle.xaml
@@ -1,126 +1,0 @@
-<!--
-    Copyright (c) Microsoft Corporation. All rights reserved. Licensed under
-    the MIT License. See LICENSE in the project root for license information.
--->
-<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                    xmlns:local="using:Microsoft.Terminal.Settings.Editor"
-                    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-                    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
-                    mc:Ignorable="d">
-
-    <ResourceDictionary.ThemeDictionaries>
-        <ResourceDictionary x:Key="Light">
-            <Style x:Key="SecondaryTextBlockStyle"
-                   TargetType="TextBlock">
-                <Setter Property="Foreground" Value="{ThemeResource TextFillColorSecondary}" />
-            </Style>
-            <SolidColorBrush x:Key="SubgroupHeaderBrush"
-                             Color="{StaticResource TextFillColorSecondary}" />
-
-            <StaticResource x:Key="SettingContainerErrorSeverityBackgroundBrush"
-                            ResourceKey="SystemFillColorCriticalBackgroundBrush" />
-            <StaticResource x:Key="SettingContainerWarningSeverityBackgroundBrush"
-                            ResourceKey="SystemFillColorCautionBackgroundBrush" />
-
-            <StaticResource x:Key="SettingContainerErrorSeverityIconBackground"
-                            ResourceKey="SystemFillColorCriticalBrush" />
-            <StaticResource x:Key="SettingContainerWarningSeverityIconBackground"
-                            ResourceKey="SystemFillColorCautionBrush" />
-
-            <StaticResource x:Key="SettingContainerResetButtonIconForeground"
-                            ResourceKey="SystemAccentColorDark2" />
-        </ResourceDictionary>
-        <ResourceDictionary x:Key="HighContrast">
-            <Style x:Key="SecondaryTextBlockStyle"
-                   TargetType="TextBlock" />
-            <SolidColorBrush x:Key="SubgroupHeaderBrush"
-                             Color="{ThemeResource SystemColorWindowTextColor}" />
-
-            <StaticResource x:Key="SettingContainerErrorSeverityBackgroundBrush"
-                            ResourceKey="SystemColorWindowColorBrush" />
-            <StaticResource x:Key="SettingContainerWarningSeverityBackgroundBrush"
-                            ResourceKey="SystemColorWindowColorBrush" />
-
-            <StaticResource x:Key="SettingContainerErrorSeverityIconBackground"
-                            ResourceKey="SystemColorHighlightColorBrush" />
-            <StaticResource x:Key="SettingContainerWarningSeverityIconBackground"
-                            ResourceKey="SystemColorHighlightColorBrush" />
-
-            <StaticResource x:Key="SettingContainerResetButtonIconForeground"
-                            ResourceKey="SystemAccentColorLight1" />
-        </ResourceDictionary>
-        <ResourceDictionary x:Key="Dark">
-            <Style x:Key="SecondaryTextBlockStyle"
-                   TargetType="TextBlock">
-                <Setter Property="Foreground" Value="{ThemeResource TextFillColorSecondary}" />
-            </Style>
-            <SolidColorBrush x:Key="SubgroupHeaderBrush"
-                             Color="{StaticResource TextFillColorSecondary}" />
-
-            <StaticResource x:Key="SettingContainerErrorSeverityBackgroundBrush"
-                            ResourceKey="SystemFillColorCriticalBackgroundBrush" />
-            <StaticResource x:Key="SettingContainerWarningSeverityBackgroundBrush"
-                            ResourceKey="SystemFillColorCautionBackgroundBrush" />
-
-            <StaticResource x:Key="SettingContainerErrorSeverityIconBackground"
-                            ResourceKey="SystemFillColorCriticalBrush" />
-            <StaticResource x:Key="SettingContainerWarningSeverityIconBackground"
-                            ResourceKey="SystemFillColorCautionBrush" />
-
-            <StaticResource x:Key="SettingContainerResetButtonIconForeground"
-                            ResourceKey="SystemAccentColorLight2" />
-        </ResourceDictionary>
-    </ResourceDictionary.ThemeDictionaries>
-
-    <x:String x:Key="SettingContainerErrorIconGlyph">&#xF13D;</x:String>
-    <x:String x:Key="SettingContainerWarningIconGlyph">&#xF13C;</x:String>
-
-    <x:Double x:Key="SettingContainerIconFontSize">16</x:Double>
-
-    <Style x:Key="SettingContainerResetButtonStyle"
-           BasedOn="{StaticResource DefaultButtonStyle}"
-           TargetType="Button">
-        <Setter Property="Margin" Value="4,0,0,0" />
-        <Setter Property="Height" Value="20" />
-        <Setter Property="Width" Value="20" />
-        <Setter Property="Padding" Value="0" />
-        <Setter Property="BorderBrush" Value="Transparent" />
-        <Setter Property="Background" Value="Transparent" />
-    </Style>
-
-    <Style x:Key="SettingContainerFontIconStyle"
-           TargetType="FontIcon">
-        <Setter Property="Foreground" Value="{ThemeResource SettingContainerResetButtonIconForeground}" />
-        <Setter Property="FontSize" Value="12" />
-        <Setter Property="FontFamily" Value="{ThemeResource SymbolThemeFontFamily}" />
-    </Style>
-
-    <Style x:Key="SettingsPageItemHeaderStyle"
-           BasedOn="{StaticResource BodyTextBlockStyle}"
-           TargetType="TextBlock">
-        <Setter Property="LineHeight" Value="20" />
-        <Setter Property="TextWrapping" Value="WrapWholeWords" />
-    </Style>
-
-    <Style x:Key="SettingsPageItemDescriptionStyle"
-           BasedOn="{StaticResource CaptionTextBlockStyle}"
-           TargetType="TextBlock">
-        <Setter Property="Margin" Value="0,0,24,0" />
-        <Setter Property="LineHeight" Value="16" />
-        <Setter Property="Foreground" Value="{ThemeResource SubgroupHeaderBrush}" />
-        <Setter Property="TextWrapping" Value="WrapWholeWords" />
-        <Setter Property="FontFamily" Value="Segoe UI, Segoe Fluent Icons, Segoe MDL2 Assets" />
-    </Style>
-
-    <Style x:Key="SettingContainerCurrentValueTextBlockStyle"
-           BasedOn="{StaticResource SettingsPageItemDescriptionStyle}"
-           TargetType="TextBlock">
-        <Setter Property="MaxWidth" Value="248" />
-        <Setter Property="Margin" Value="0" />
-        <Setter Property="HorizontalAlignment" Value="Right" />
-        <Setter Property="VerticalAlignment" Value="Center" />
-        <Setter Property="FontFamily" Value="Segoe UI, Segoe Fluent Icons, Segoe MDL2 Assets" />
-    </Style>
-</ResourceDictionary>
