--- src/cascadia/TerminalSettingsEditor/StyleExtensions.idl
@@ -1,28 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-namespace Microsoft.Terminal.Settings.Editor
-{
-    // Attached property that lets a Style merge an extra ResourceDictionary into
-    // the target FrameworkElement's Resources.MergedDictionaries. Used by
-    // SettingsCard / SettingsExpander to inject right-aligned ToggleSwitch and
-    // sized Slider / ComboBox / TextBox defaults so children laid out inside a
-    // card look right out of the box. Mirrors the Windows Community Toolkit's
-    // CommunityToolkit.WinUI.Controls.StyleExtensions.
-    static runtimeclass StyleExtensions
-    {
-        static Windows.UI.Xaml.DependencyProperty ResourcesProperty { get; };
-
-        static Windows.UI.Xaml.ResourceDictionary GetResources(Windows.UI.Xaml.DependencyObject target);
-        static void SetResources(Windows.UI.Xaml.DependencyObject target, Windows.UI.Xaml.ResourceDictionary value);
-
-        // Idempotently merges the project's SettingsControlsImplicitStyles.xaml
-        // dictionary into the target's Resources.MergedDictionaries. Called from
-        // SettingsCard / SettingsExpander OnApplyTemplate to give children
-        // (ToggleSwitch / Slider / ComboBox / TextBox) sensible Windows 11
-        // defaults without each call site setting Style explicitly. Loaded from
-        // C++ rather than via the attached DP because the WCT Setter.Value
-        // pattern crashes WinUI 2 in this codebase.
-        static void EnsureImplicitStylesMergedInto(Windows.UI.Xaml.FrameworkElement target);
-    }
-}
